app/requirements.txt
Flask
flask-restful
marshmallow
psycopg2-binary
redis